Platform and Unzip

+++ index.py
@@ -1,7 +1,8 @@
 #!/usr/bin/python3
+import platform
 import subprocess, gdown, os.path
 from getpass import getpass
-import platform
+from zipfile import ZipFile
 
 print('Vos informations GITLAB de connexion vous seront demandées.')
 username = input('User: ')
@@ -10,11 +11,11 @@ password = getpass(prompt='Pass: ')
 url_addons = "git clone https://{}:{}@gitlab.com/dev-odoo-14/addons.git".format(
     username, password)
 url_docker_compose = "git clone https://github.com/barryab12/openmoise-docker.git"
-filestore_link = "https://drive.google.com/uc?id=1jMHuiniqvX11xJAlUzZtHvZM0eJPHnFk"
+id = "1QcQeOyvFhqQ0dlLWX5_lySdgnyzum0BT"
 
 print("-- Téléchargement des images et fichiers documents ... ")
 if not os.path.isdir('filestore/'):
-    gdown.download(filestore_link, 'filestore.zip', quiet=False)
+    gdown.download(id=id, output='filestore.zip', quiet=False)
 
 print("-- Téléchargement des modules")
 if not os.path.isdir('addons/'):
@@ -25,9 +26,13 @@ subprocess.run(url_docker_compose, shell=True)
 subprocess.run('mv openmoise-docker/* .', shell=True)
 subprocess.run('rm -fr openmoise-docker/', shell=True)
 if not os.path.isfile('filestore.zip'):
-    subprocess.run('unzip filestore.zip', shell=True)
+    with ZipFile('filestore.zip', 'r') as file:
+        file.extractall()
 print("-- Suppression des fichiers ... ")
-subprocess.run('rm -fr filestore.zip', shell=True)
+if (platform.system() == "Linux"):
+    subprocess.run('rm -fr filestore.zip', shell=True)
+elif (platform.system() == "Windows"):
+    subprocess.run('del filestore.zip', shell=True)
 print("-- Lancement des conteneurs")
 if os.path.isfile('docker-compose.yml') or os.path.isfile(
         'docker-compose.yaml'):
